add the diag refactor
Refactor the diag commands to fit the new code lib/ code structure.
Each version of the diag function which has it's own gen specific
functionalities will have a function defined for it in the /genX
library folder. Add the functional pointers as well.
