+++ lib/gen_ops.h
@@ -41,22 +41,22 @@
  */
 const struct switchtec_gen_ops switchtec_gen6_ops = {
 	/* Diagnostics - NULL means use common implementation or not supported */
-	.diag_cross_hair_enable = NULL,
-	.diag_cross_hair_disable = NULL,
-	.diag_cross_hair_get = NULL,
+	.diag_cross_hair_enable = switchtec_diag_cross_hair_enable_gen4,
+	.diag_cross_hair_disable = switchtec_diag_cross_hair_disable_gen4,
+	.diag_cross_hair_get = switchtec_diag_cross_hair_get_gen4,
 	.diag_eye_set_mode = NULL,
-	.diag_eye_start = NULL,
+	.diag_eye_start = switchtec_diag_eye_start_gen6,
 	.diag_eye_fetch = NULL,
 	.diag_eye_cancel = NULL,
 	.diag_eye_read = NULL,
-	.diag_loopback_set = NULL,
-	.diag_loopback_get = NULL,
-	.diag_pattern_gen_set = NULL,
-	.diag_pattern_gen_get = NULL,
-	.diag_pattern_mon_set = NULL,
-	.diag_pattern_mon_get = NULL,
-	.diag_pattern_inject = NULL,
-	.diag_ltssm_log = NULL,
+	.diag_loopback_set = switchtec_diag_loopback_set_gen5,
+	.diag_loopback_get = switchtec_diag_loopback_get_gen5,
+	.diag_pattern_gen_set = switchtec_diag_pattern_gen_set_gen6,
+	.diag_pattern_gen_get = switchtec_diag_pattern_gen_get_gen6,
+	.diag_pattern_mon_set = switchtec_diag_pattern_mon_set_gen6,
+	.diag_pattern_mon_get = switchtec_diag_pattern_mon_get_gen6,
+	.diag_pattern_inject = switchtec_diag_pattern_inject_gen4,
+	.diag_ltssm_log = switchtec_diag_ltssm_log_gen6,
 	.diag_ltssm_log_set = NULL,
 	.diag_port_eq_tx_coeff = NULL,
 	.diag_port_eq_tx_table = NULL,
@@ -64,14 +64,21 @@ const struct switchtec_gen_ops switchtec_gen6_ops = {
 	.diag_rcvr_obj = NULL,
 	.diag_rcvr_ext = NULL,
 	.diag_refclk_ctl = NULL,
-	.inject_err_tlp_lcrc = NULL,
+	.inject_err_tlp_lcrc = switchtec_inject_err_tlp_lcrc_gen5,
 	.inject_err_cto = NULL,
-	.inject_err_ack_nack = NULL,
-	.inject_err_tlp_seqnum = NULL,
+	.inject_err_ack_nack = switchtec_inject_err_ack_nack_gen4,
+	.inject_err_tlp_seqnum = switchtec_inject_err_tlp_seqnum_gen4,
 	.inject_err_tlp_ecrc = NULL,
-	.inject_err_dllp_crc = NULL,
+	.inject_err_dllp_crc = switchtec_inject_err_dllp_crc_gen4,
 	.inject_err_dllp = NULL,
 	.inject_err_dup_tlp = NULL,
+	.osa_capture_data = switchtec_osa_capture_data_gen6,
+	.osa_capture_control = switchtec_osa_capture_control_gen6,
+	.osa_config_misc = switchtec_osa_config_misc_gen5,
+	.osa_config_pattern = switchtec_osa_config_pattern_gen5,
+	.osa_config_type = switchtec_osa_config_type_gen5,
+	.osa_dump_conf = switchtec_osa_dump_conf_gen6,
+	.osa = switchtec_osa_gen6,
 
 	/* Manufacturing */
 	.security_config_get = NULL,
@@ -121,36 +128,43 @@ const struct switchtec_gen_ops switchtec_gen6_ops = {
  */
 const struct switchtec_gen_ops switchtec_gen5_ops = {
 	/* Diagnostics - NULL means use common implementation or not supported */
-	.diag_cross_hair_enable = NULL,
-	.diag_cross_hair_disable = NULL,
-	.diag_cross_hair_get = NULL,
+	.diag_cross_hair_enable = switchtec_diag_cross_hair_enable_gen4,
+	.diag_cross_hair_disable = switchtec_diag_cross_hair_disable_gen4,
+	.diag_cross_hair_get = switchtec_diag_cross_hair_get_gen4,
 	.diag_eye_set_mode = NULL,
-	.diag_eye_start = NULL,
+	.diag_eye_start = switchtec_diag_eye_start_gen5,
 	.diag_eye_fetch = NULL,
 	.diag_eye_cancel = NULL,
-	.diag_eye_read = NULL,
-	.diag_loopback_set = NULL,
-	.diag_loopback_get = NULL,
-	.diag_pattern_gen_set = NULL,
-	.diag_pattern_gen_get = NULL,
-	.diag_pattern_mon_set = NULL,
-	.diag_pattern_mon_get = NULL,
-	.diag_pattern_inject = NULL,
-	.diag_ltssm_log = NULL,
+	.diag_eye_read = switchtec_diag_eye_read_gen5,
+	.diag_loopback_set = switchtec_diag_loopback_set_gen5,
+	.diag_loopback_get = switchtec_diag_loopback_get_gen5,
+	.diag_pattern_gen_set = switchtec_diag_pattern_gen_set_gen5,
+	.diag_pattern_gen_get = switchtec_diag_pattern_gen_get_gen4,
+	.diag_pattern_mon_set = switchtec_diag_pattern_mon_set_gen4,
+	.diag_pattern_mon_get = switchtec_diag_pattern_mon_get_gen4,
+	.diag_pattern_inject = switchtec_diag_pattern_inject_gen4,
+	.diag_ltssm_log = switchtec_diag_ltssm_log_gen5,
 	.diag_ltssm_log_set = NULL,
-	.diag_port_eq_tx_coeff = NULL,
-	.diag_port_eq_tx_table = NULL,
-	.diag_port_eq_tx_fslf = NULL,
+	.diag_port_eq_tx_coeff = switchtec_diag_port_eq_tx_coeff_gen5,
+	.diag_port_eq_tx_table = switchtec_diag_port_eq_tx_table_gen5,
+	.diag_port_eq_tx_fslf = switchtec_diag_port_eq_tx_fslf_gen5,
 	.diag_rcvr_obj = NULL,
 	.diag_rcvr_ext = NULL,
 	.diag_refclk_ctl = NULL,
-	.inject_err_tlp_lcrc = NULL,
-	.inject_err_cto = NULL,
-	.inject_err_ack_nack = NULL,
-	.inject_err_tlp_seqnum = NULL,
+	.inject_err_tlp_lcrc = switchtec_inject_err_tlp_lcrc_gen5,
+	.inject_err_cto = switchtec_inject_err_cto_gen5,
+	.inject_err_ack_nack = switchtec_inject_err_ack_nack_gen4,
+	.inject_err_tlp_seqnum = switchtec_inject_err_tlp_seqnum_gen4,
 	.inject_err_tlp_ecrc = NULL,
-	.inject_err_dllp_crc = NULL,
+	.inject_err_dllp_crc = switchtec_inject_err_dllp_crc_gen4,
 	.inject_err_dup_tlp = NULL,
+	.osa_capture_data = switchtec_osa_capture_data_gen5,
+	.osa_capture_control = switchtec_osa_capture_control_gen5,
+	.osa_config_misc = switchtec_osa_config_misc_gen5,
+	.osa_config_pattern = switchtec_osa_config_pattern_gen5,
+	.osa_config_type = switchtec_osa_config_type_gen5,
+	.osa_dump_conf = switchtec_osa_dump_conf_gen5,
+	.osa = switchtec_osa_gen5,
 
 	/* Manufacturing */
 	.security_config_get = NULL,
@@ -199,37 +213,44 @@ const struct switchtec_gen_ops switchtec_gen5_ops = {
  */
 const struct switchtec_gen_ops switchtec_gen4_ops = {
 	/* Diagnostics - NULL means use common implementation or not supported */
-	.diag_cross_hair_enable = NULL,
-	.diag_cross_hair_disable = NULL,
-	.diag_cross_hair_get = NULL,
-	.diag_eye_set_mode = NULL,
-	.diag_eye_start = NULL,
-	.diag_eye_fetch = NULL,
-	.diag_eye_cancel = NULL,
-	.diag_eye_read = NULL,
-	.diag_loopback_set = NULL,
-	.diag_loopback_get = NULL,
-	.diag_pattern_gen_set = NULL,
-	.diag_pattern_gen_get = NULL,
-	.diag_pattern_mon_set = NULL,
-	.diag_pattern_mon_get = NULL,
-	.diag_pattern_inject = NULL,
-	.diag_ltssm_log = NULL,
+	.diag_cross_hair_enable = switchtec_diag_cross_hair_enable_gen4,
+	.diag_cross_hair_disable = switchtec_diag_cross_hair_disable_gen4,
+	.diag_cross_hair_get = switchtec_diag_cross_hair_get_gen4,
+	.diag_eye_set_mode = switchtec_diag_eye_set_mode_gen4,
+	.diag_eye_start = switchtec_diag_eye_start_gen5,
+	.diag_eye_fetch = switchtec_diag_eye_fetch_gen4,
+	.diag_eye_cancel = switchtec_diag_eye_cancel_gen4,
+	.diag_eye_read = switchtec_diag_eye_read_gen5,
+	.diag_loopback_set = switchtec_diag_loopback_set_gen5,
+	.diag_loopback_get = switchtec_diag_loopback_get_gen5,
+	.diag_pattern_gen_set = switchtec_diag_pattern_gen_set_gen5,
+	.diag_pattern_gen_get = switchtec_diag_pattern_gen_get_gen4,
+	.diag_pattern_mon_set = switchtec_diag_pattern_mon_set_gen4,
+	.diag_pattern_mon_get = switchtec_diag_pattern_mon_get_gen4,
+	.diag_pattern_inject = switchtec_diag_pattern_inject_gen4,
+	.diag_ltssm_log = switchtec_diag_ltssm_log_gen5,
 	.diag_ltssm_log_set = NULL,
-	.diag_port_eq_tx_coeff = NULL,
-	.diag_port_eq_tx_table = NULL,
-	.diag_port_eq_tx_fslf = NULL,
-	.diag_rcvr_obj = NULL,
+	.diag_port_eq_tx_coeff = switchtec_diag_port_eq_tx_coeff_gen5,
+	.diag_port_eq_tx_table = switchtec_diag_port_eq_tx_table_gen5,
+	.diag_port_eq_tx_fslf = switchtec_diag_port_eq_tx_fslf_gen5,
+	.diag_rcvr_obj = switchtec_diag_rcvr_obj_gen4,
 	.diag_rcvr_ext = NULL,
 	.diag_refclk_ctl = NULL,
-	.inject_err_tlp_lcrc = NULL,
-	.inject_err_cto = NULL,
-	.inject_err_ack_nack = NULL,
-	.inject_err_tlp_seqnum = NULL,
+	.inject_err_tlp_lcrc = switchtec_inject_err_tlp_lcrc_gen5,
+	.inject_err_cto = switchtec_inject_err_cto_gen5,
+	.inject_err_ack_nack = switchtec_inject_err_ack_nack_gen4,
+	.inject_err_tlp_seqnum = switchtec_inject_err_tlp_seqnum_gen4,
 	.inject_err_tlp_ecrc = NULL,
-	.inject_err_dllp_crc = NULL,
-	.inject_err_dllp = NULL,
+	.inject_err_dllp_crc = switchtec_inject_err_dllp_crc_gen4,
+	.inject_err_dllp = switchtec_inject_err_dllp_gen4,
 	.inject_err_dup_tlp = NULL,
+	.osa_capture_data = NULL,
+	.osa_capture_control = NULL,
+	.osa_config_misc = NULL,
+	.osa_config_pattern = NULL,
+	.osa_config_type = NULL,
+	.osa_dump_conf = NULL,
+	.osa = NULL,
 
 	/* Manufacturing */
 	.security_config_get = NULL,
@@ -290,8 +311,8 @@ const struct switchtec_gen_ops switchtec_gen3_ops = {
 	.diag_eye_fetch = NULL,
 	.diag_eye_cancel = NULL,
 	.diag_eye_read = NULL,
-	.diag_loopback_set = NULL,
-	.diag_loopback_get = NULL,
+	.diag_loopback_set = switchtec_diag_loopback_set_gen4,
+	.diag_loopback_get = switchtec_diag_loopback_get_gen4,
 	.diag_pattern_gen_set = NULL,
 	.diag_pattern_gen_get = NULL,
 	.diag_pattern_mon_set = NULL,
@@ -299,9 +320,9 @@ const struct switchtec_gen_ops switchtec_gen3_ops = {
 	.diag_pattern_inject = NULL,
 	.diag_ltssm_log = NULL, //Not supported
 	.diag_ltssm_log_set = NULL, //Not supported
-	.diag_port_eq_tx_coeff = NULL,
-	.diag_port_eq_tx_table = NULL,
-	.diag_port_eq_tx_fslf = NULL,
+	.diag_port_eq_tx_coeff = switchtec_diag_port_eq_tx_coeff_gen4,
+	.diag_port_eq_tx_table = switchtec_diag_port_eq_tx_table_gen4,
+	.diag_port_eq_tx_fslf = switchtec_diag_port_eq_tx_fslf_gen4,
 	.diag_rcvr_obj = NULL,
 	.diag_rcvr_ext = NULL,
 	.diag_refclk_ctl = NULL,
@@ -312,6 +333,13 @@ const struct switchtec_gen_ops switchtec_gen3_ops = {
 	.inject_err_tlp_ecrc = NULL,
 	.inject_err_dllp_crc = NULL,
 	.inject_err_dup_tlp = NULL,
+	.osa_capture_data = NULL,
+	.osa_capture_control = NULL,
+	.osa_config_misc = NULL,
+	.osa_config_pattern = NULL,
+	.osa_config_type = NULL,
+	.osa_dump_conf = NULL,
+	.osa = NULL,
 
 	/* Manufacturing - Gen3 uses different structures */
 	.security_config_get = NULL,
